refactored extension.ts into separate files- one for comments and one for findCapitalizedPrimitiveTypes
